fix for demo
audio and video in realtime, fix audio glitches, fix odas positions

+++ steno/src/model/classifier/classifier.cpp
@@ -5,7 +5,7 @@
 namespace Model
 {
 std::vector<int> Classifier::classify(const std::vector<SourcePosition> &audioPositions,
-                                      const std::vector<SphericalAngleRect> &imagePositions, const int &rangeThreshold)
+                                      const std::vector<SphericalAngleRect> &imagePositions, const float &rangeThreshold)
 {
     std::vector<int> sourcesToSuppress;
