+++ steno/src/model/classifier/classifier.h
@@ -12,7 +12,7 @@ class Classifier
 {
    public:
     static std::vector<int> classify(const std::vector<SourcePosition> &audioPositions,
-                                     const std::vector<SphericalAngleRect> &imagePositions, const int &rangeThreshold);
+                                     const std::vector<SphericalAngleRect> &imagePositions, const float &rangeThreshold);
 };
 
 }    // namespace Model
